issue: Limit Ticket Access Via URL
This addresses issue 4321 where an Agent is able to change the `id=` param
for `scp/tickets.php` and get access to tickets not in their Department.
They can only post Internal Notes and all options like Transfer, Assign,
etc. are Disabled but they still get View access though. So this updates
the `checkStaffPerm()` check to fail if the Agent does not have access to
the Department the ticket is under.