feature: Answered Role Permission
This feature adds a new Role permission called Answer. This permission (if
Enabled) allows an Agent to mark tickets as answered or unanswered. This
permission is configurable for each and every Role in the helpdesk. This
also keeps current functionality where Department Managers do not need the
Role Permission in order to mark tickets as answered/unaswered. In addition,
this adds a "Mark as Answered" checkbox next to the ticket status dropdown
on Internal Notes so that Agents may quickly mark the ticket as answered
when posting an internal note.