issue: Queue Column Conditions
This addresses an issue where Queue Column conditions are working in the
preview but not in the actual queue view. This is due to the `getQueue()`
function that returns the CustomQueue object. In the preview it can find the
object just fine and return it. In the ticket view however, it cannot find
the object and returns null. This updates the function to lookup the
CustomQueue by `queue_id` if there is no CustomQueue object found.