issue: Recipient Link Canned Responses
This addresses issue 1459 where `%{recipient.ticket_link}` is not being
replaced when used in a Canned Response. If you use all the other defined
Redactor links in a Canned Response they work fine, only the recipient's
ticket link is broken. This is due to the `recipient` object not being
available in the variable replacer for canned responses. This adds the
`recipient` object to the variable replacer which replaces
`%{recipient.ticket_link}` with the recipient's actual link.